Created Navigation Abstraction
