+++ app/src/main/java/com/example/b07group7project/Navigation.java
@@ -1,0 +1,17 @@
+package com.example.b07group7project;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentTransaction;
+
+public abstract class Navigation extends AppCompatActivity implements FragmentSwitch {
+
+    @Override
+    public void replaceFragment(Fragment newFragment) {
+        FragmentManager fragmentManager = getSupportFragmentManager();
+        FragmentTransaction fragmentTransaction = fragmentManager.beginTransaction();
+        fragmentTransaction.replace(R.id.frameLayout, newFragment);
+        fragmentTransaction.commit();
+    }
+}
