+++ app/src/main/java/com/example/b07group7project/ShopperNavigationActivity.java
@@ -1,15 +1,12 @@
 package com.example.b07group7project;
 
-import androidx.appcompat.app.AppCompatActivity;
 import androidx.fragment.app.Fragment;
-import androidx.fragment.app.FragmentManager;
-import androidx.fragment.app.FragmentTransaction;
 
 import android.os.Bundle;
 
 import com.example.b07group7project.databinding.ShopperNavigationActivityBinding;
 
-public class ShopperNavigationActivity extends AppCompatActivity implements FragmentSwitch {
+public class ShopperNavigationActivity extends Navigation {
 
     ShopperNavigationActivityBinding binding;
 
@@ -44,13 +41,5 @@ public class ShopperNavigationActivity extends AppCompatActivity implements Frag
 
             return true;
         });
-    }
-
-    @Override
-    public void replaceFragment(Fragment newFragment) {
-        FragmentManager fragmentManager = getSupportFragmentManager();
-        FragmentTransaction fragmentTransaction = fragmentManager.beginTransaction();
-        fragmentTransaction.replace(R.id.frameLayout, newFragment);
-        fragmentTransaction.commit();
     }
 }
